Set up document processing

+++ database.py
@@ -1,20 +1,20 @@
+import time
 import haystack.nodes
 from haystack.document_stores.faiss import FAISSDocumentStore
 from haystack.nodes import PDFToTextConverter, PDFToTextOCRConverter, TextConverter, FileTypeClassifier
-from haystack.nodes import PreProcessor, DocxToTextConverter, EmbeddingRetriever
-from engine_v2 import Engine
+from haystack.nodes import PreProcessor, DocxToTextConverter
 import logging
 from haystack.pipelines import Pipeline
 import os
-#from special_reports_extractor import SRExtractor
-import re
+from special_reports_extractor import SrExtractor
 
-haystack.HAYSRACK_TELEMETRY_ENABLED = False
+
+haystack.HAYSTACK_TELEMETRY_ENABLED = False
+
 
 class Database:
 
     def __init__(self):
-        # will need change
         self.document_store = None
         self.file_classifier = None
         self.converter_pdf = None
@@ -22,26 +22,28 @@ class Database:
         self.converter_text = None
         self.pre_processor = None
         self.extractor = None
+        self.current_number_of_documents = 0
+        self.name = ""
 
-    @staticmethod
-    def load_document_store(faiss_index_path: str, faiss_config_path: str):
-        database = Database()
-        database.document_store = FAISSDocumentStore.load(index_path=faiss_index_path, config_path=faiss_config_path)
+    def load_document_store(self, faiss_index_path: str, faiss_config_path: str):
+
+        self.document_store = FAISSDocumentStore.load(index_path=faiss_index_path, config_path=faiss_config_path)
         logging.info("Database loaded. Existing Document Store initialized.")
-        return database
 
     def update_document_store(self, path_to_documents: str, retriever: haystack.nodes.EmbeddingRetriever):
+
         assert self.document_store is not None, "No Document Store to update. " \
                                             "Please load a database or create a new one"
         self.file_classifier = FileTypeClassifier()
         self.converter_pdf = PDFToTextOCRConverter()
         self.converter_docx = DocxToTextConverter()
         self.converter_text = TextConverter()
+        self.extractor = SrExtractor()
         self.pre_processor = PreProcessor(clean_empty_lines=True,
                                           clean_whitespace=True,
                                           clean_header_footer=True,
                                           split_by="passage",
-                                          split_respect_sentence_boundary=True,
+                                          split_respect_sentence_boundary=False,
                                           split_overlap=0)
 
         tray_dir = os.listdir(path_to_documents)
@@ -49,25 +51,27 @@ class Database:
         new_documents = self.load_documents(path_to_documents)
         self.document_store.write_documents(new_documents)
         # add assertion, must be the same retriever
-        self.document_store.update_embeddings(retriever=retriever, update_existing_embeddings=False)
+        self.document_store.update_embeddings(retriever=retriever, update_existing_embeddings=True)
         logging.info("New documents embedded. Database updated.")
-
         """
-        for f in os.listdir(tray_dir):
-            if not f.endswith(".txt") or f.endswith(".pdf") or f.endswith(".docx"):
-                continue
-            assert os.chdir == os.path.join(tray_dir), "Wrong directory!"
-            os.remove(os.path.join(tray_dir, f))
+        if not test:
+            for f in os.listdir(tray_dir):
+                if not f.endswith(".txt") or f.endswith(".pdf") or f.endswith(".docx"):
+                    continue
+                assert os.chdir == os.path(tray_dir), "Wrong directory!"
+                os.remove(os.path.join(tray_dir, f))
         """
 
-    def launch_new_document_store(self):
+    def launch_new_document_store(self, set_file_name="special_reports_faiss_store"):
+        print(os.getcwd())
+        assert not os.path.exists(f"{os.getcwd()}/{set_file_name}.db"), "A Document Store with this name " \
+                                                                        "already exits. Please load it."
+
         self.document_store = FAISSDocumentStore(faiss_index_factory_str="Flat",
-                                                 sql_url="sqlite:///special_reports_faiss_store.db",
-                                                 embedding_dim=764,
+                                                 sql_url=f"sqlite:///{set_file_name}.db",
+                                                 embedding_dim=768,
                                                  similarity="dot_product")
         logging.info("Database created. Initialized a new Document Store")
-
-
 
     def load_documents(self, path_to_documents: str):
         """This method take a relative path to a folder with documents of different types.
@@ -87,32 +91,25 @@ class Database:
                                         inputs=["FileTypeClassifier.output_2"])
         pipeline_preprocessing.add_node(component=self.converter_docx, name="DocxConverter",
                                         inputs=["FileTypeClassifier.output_4"])
-        """
-        pipeline_preprocessing.add_node(component=self.extractor, name="SRExtractor",
+        pipeline_preprocessing.add_node(component=self.extractor, name="SrExtractor",
                                         inputs=["TextConverter", "PdfConverter", "DocxConverter"])
-        """
+
         for file in os.listdir(path_to_documents):
-            # 1.open path_to_documents and open one-by-one, classify the type, convert to text
+            # Open path_to_documents and open one-by-one, classify the type,
+            # convert to text, extract metadata and paragraphs
             converted_document = pipeline_preprocessing.run(file_paths=[os.path.join(path_to_documents, file)])
-            # 2. Extract metadata for the whole document
-            text_raw = self.document_preprocess_for_extration(converted_document["documents"][0].content)
-            report_title, report_info = self.extract_metadata(text_raw)
-            # 3. Split into paragraphs and save paragraph numbers in a list
-            paragraph_numbers, paragraphs = self.extract_paragraphs(text_raw)
-            # 4. Pre-process each paragraph, save as Documents, and metadata to each
-            document_paragraphs = self.pre_processor.process(paragraphs)
-            assert len(document_paragraphs) == len(paragraph_numbers), "Too many documents"
+            documents_processed.append(self.pre_processor.process(converted_document["documents"]))
 
-            # add metadata and paragraph number to each paragraph
-            i = 0
-            for doc in document_paragraphs:
-                doc.meta["report_title"] = report_title
-                doc.meta["report_info"] = report_info
-                doc.meta["paragraph_number"] = paragraph_numbers[i]
-                i += 1
+        self.current_number_of_documents += len(documents_processed[0])
 
-            documents_processed.append(document_paragraphs)
+        logging.info("Database update successfully.")
 
-        return documents_processed
+        return documents_processed[0]
 
+    def save_database(self, path: str):
+        """This method saves the Document Store of the database with a time stamp."""
 
+        self.name = f"special_report_database_at_{str(int(time.time())).replace(' ', '_')}"
+        self.document_store.save(index_path=os.path.join(path, self.name + ".faiss"))
+
+        logging.info("Database saved successfully.")
